ADD: email -> confirm -> guestList -> acknowledge
