+++ src/app/api/volunteers/data/route.ts
@@ -13,19 +13,21 @@ export async function GET() {
   }
 }
 
-// TODO: maybe this should be its own route, rather than a weird PUT/POST distinction
+// Update the "acknowledged" field in storage
 export async function POST(request: Request) {
   const entry = await request.json();
 
   try {
-    const entry = await request.json();
     const data = await FormSubmissions.markEntryAcknowledged(entry);
+
+    // console.log({ data })
 
     return NextResponse.json(
       { message: "volunteer acknowledged email confirmation", data },
       { status: 200 },
     );
   } catch (err) {
+    // console.log({ err })
     return NextResponse.json(
       {
         message: `error when acknowledging email receipt for ${entry.name} re: ${entry.date}`,
@@ -36,12 +38,16 @@ export async function POST(request: Request) {
   }
 }
 
+// Manually update the "confirmed?" field in storage
 export async function PUT(request: Request) {
   // TODO: sanitize input here or on FE?
   const entry = await request.json();
   try {
     const data = await FormSubmissions.updateEntryConfirmed(entry);
-    return NextResponse.json({ message: "response!", data }, { status: 200 });
+    return NextResponse.json(
+      { message: "Successfully updated confirmed status!", data },
+      { status: 200 },
+    );
   } catch (err) {
     return NextResponse.json(
       {
